Extract CLI helpers into cli.py to untangle the import graph
normalize.py and extract_comment.py imported parse_args, set_logging
and add_default_arguments from __main__, putting the entry point in the
middle of the dependency graph. This forced the previous commit to move
the aliases serializer into create_content.py to avoid a circular
import.

With the CLI helpers in their own module, __main__ is back on top of
the graph and dump_normalized_aliases/get_new_aliases return to
normalize.py where they belong.
